Fixed tests in msal-common, working on msal-browser
