+++ lib/msal-browser/test/cache/BrowserStorage.spec.ts
@@ -5,8 +5,8 @@ import { BrowserStorage } from "../../src/cache/BrowserStorage";
 import { TEST_CONFIG, TEST_TOKENS, TEST_DATA_CLIENT_INFO, RANDOM_TEST_GUID } from "../utils/StringConstants";
 import { CacheOptions } from "../../src/config/Configuration";
 import { BrowserConfigurationAuthErrorMessage, BrowserConfigurationAuthError } from "../../src/error/BrowserConfigurationAuthError";
-import { ICacheStorage, Constants, PersistentCacheKeys, TemporaryCacheKeys } from "@azure/msal-common";
-import { BrowserConstants } from "../../src/utils/BrowserConstants";
+import { ICacheStorage, Constants, PersistentCacheKeys, InMemoryCache } from "@azure/msal-common";
+import { BrowserConstants, TemporaryCacheKeys } from "../../src/utils/BrowserConstants";
 
 class TestCacheStorage implements ICacheStorage {
     setItem(key: string, value: string): void {
@@ -26,7 +26,13 @@ class TestCacheStorage implements ICacheStorage {
     }
     clear(): void {
         throw new Error("Method not implemented.");
-    }
+	}
+	getCache(): InMemoryCache {
+		return null;
+	}
+	setCache(): InMemoryCache {
+		return null;
+	}
 }
 
 describe("BrowserStorage() tests", () => {
@@ -102,8 +108,8 @@ describe("BrowserStorage() tests", () => {
             expect(browserStorage.getItem(PersistentCacheKeys.ID_TOKEN)).to.be.eq(TEST_TOKENS.IDTOKEN_V2);
             expect(browserStorage.getItem(PersistentCacheKeys.CLIENT_INFO)).to.be.eq(TEST_DATA_CLIENT_INFO.TEST_RAW_CLIENT_INFO);
             expect(browserStorage.getItem(PersistentCacheKeys.ERROR)).to.be.eq(errorKeyVal);
-            expect(browserStorage.getItem(PersistentCacheKeys.ERROR_DESC)).to.be.eq(errorDescVal);
-        });
+			expect(browserStorage.getItem(PersistentCacheKeys.ERROR_DESC)).to.be.eq(errorDescVal);
+		});
     });
 
     describe("Interface functions", () => {
@@ -309,5 +315,98 @@ describe("BrowserStorage() tests", () => {
             const expectedDate = new Date(currentTime + (cookieLifeDays * COOKIE_LIFE_MULTIPLIER));
             expect(browserLocalStorage.getCookieExpirationTime(cookieLifeDays)).to.be.eq(expectedDate.toUTCString());
         });
-    });
+	});
+	
+	describe("Helpers", () => {
+
+		it("generateAuthorityKey() creates a valid cache key for authority strings", () => {
+			const browserStorage = new BrowserStorage(TEST_CONFIG.MSAL_CLIENT_ID, cacheConfig);
+            const authorityKey = browserStorage.generateAuthorityKey(RANDOM_TEST_GUID);
+            expect(authorityKey).to.be.eq(`${TemporaryCacheKeys.AUTHORITY}${Constants.RESOURCE_DELIM}${RANDOM_TEST_GUID}`);
+        });
+
+        it("generateNonceKey() create a valid cache key for nonce strings", () => {
+			const browserStorage = new BrowserStorage(TEST_CONFIG.MSAL_CLIENT_ID, cacheConfig);
+            const nonceKey = browserStorage.generateNonceKey(RANDOM_TEST_GUID);
+            expect(nonceKey).to.be.eq(`${TemporaryCacheKeys.NONCE_IDTOKEN}${Constants.RESOURCE_DELIM}${RANDOM_TEST_GUID}`);
+		});
+		
+		it("updateCacheEntries() correctly updates the account, authority and state in the cache", () => {
+            const authorityCacheSpy = sinon.spy(CacheHelpers.prototype, "setAuthorityCache");
+            const accountCacheSpy = sinon.spy(CacheHelpers.prototype, "setAccountCache");
+            cacheHelpers.updateCacheEntries(serverAuthParams, testAccount);
+
+            expect(accountCacheSpy.calledOnce).to.be.true;
+            expect(authorityCacheSpy.calledOnce).to.be.true;
+            const accountKey = cacheHelpers.generateAcquireTokenAccountKey(testAccount.homeAccountIdentifier);
+            const nonceKey = cacheHelpers.generateNonceKey(RANDOM_TEST_GUID);
+            const authorityKey = cacheHelpers.generateAuthorityKey(RANDOM_TEST_GUID);
+
+            expect(store[accountKey]).to.be.eq(JSON.stringify(testAccount));
+            expect(store[TemporaryCacheKeys.REQUEST_STATE]).to.be.eq(RANDOM_TEST_GUID);
+            expect(store[nonceKey]).to.be.eq(RANDOM_TEST_GUID);
+            expect(store[authorityKey]).to.be.eq(`${Constants.DEFAULT_AUTHORITY}/`);
+        });
+
+        it("updateCacheEntries() does not set account if account is given as null/empty", () => {
+            const authorityCacheSpy = sinon.spy(CacheHelpers.prototype, "setAuthorityCache");
+            const accountCacheSpy = sinon.spy(CacheHelpers.prototype, "setAccountCache");
+            cacheHelpers.updateCacheEntries(serverAuthParams, null);
+
+            expect(accountCacheSpy.calledOnce).to.be.false;
+            expect(authorityCacheSpy.calledOnce).to.be.true;
+            const accountKey = cacheHelpers.generateAcquireTokenAccountKey(testAccount.homeAccountIdentifier);
+            const nonceKey = cacheHelpers.generateNonceKey(RANDOM_TEST_GUID);
+            const authorityKey = cacheHelpers.generateAuthorityKey(RANDOM_TEST_GUID);
+
+            expect(store[accountKey]).to.be.undefined;
+            expect(store[TemporaryCacheKeys.REQUEST_STATE]).to.be.eq(RANDOM_TEST_GUID);
+            expect(store[nonceKey]).to.be.eq(RANDOM_TEST_GUID);
+            expect(store[authorityKey]).to.be.eq(`${Constants.DEFAULT_AUTHORITY}/`);
+        });
+
+        it("resetTempCacheItems() resets all temporary cache items with the given state", () => {
+            cacheHelpers.updateCacheEntries(serverAuthParams, testAccount);
+            cacheStorage.setItem(TemporaryCacheKeys.REQUEST_PARAMS, "TestRequestParams");
+            cacheStorage.setItem(TemporaryCacheKeys.ORIGIN_URI, TEST_URIS.TEST_REDIR_URI);
+
+            cacheHelpers.resetTempCacheItems(RANDOM_TEST_GUID);
+            const accountKey = cacheHelpers.generateAcquireTokenAccountKey(testAccount.homeAccountIdentifier);
+            const nonceKey = cacheHelpers.generateNonceKey(RANDOM_TEST_GUID);
+            const authorityKey = cacheHelpers.generateAuthorityKey(RANDOM_TEST_GUID);
+            expect(store[accountKey]).to.be.eq(JSON.stringify(testAccount));
+            expect(store[nonceKey]).to.be.undefined;
+            expect(store[authorityKey]).to.be.undefined;
+            expect(store[TemporaryCacheKeys.REQUEST_STATE]).to.be.undefined;
+            expect(store[TemporaryCacheKeys.REQUEST_PARAMS]).to.be.undefined;
+            expect(store[TemporaryCacheKeys.ORIGIN_URI]).to.be.undefined;
+		});
+		
+		it("Throws error if request cannot be retrieved from cache", async () => {
+			const codeRequest: AuthorizationCodeRequest = {
+				redirectUri: TEST_URIS.TEST_REDIR_URI,
+				scopes: ["scope"],
+				code: null
+			};
+			await expect(Client.acquireToken(codeRequest, RANDOM_TEST_GUID, "")).to.be.rejectedWith(ClientAuthErrorMessage.tokenRequestCacheError.desc);
+			expect(defaultAuthConfig.storageInterface.getKeys()).to.be.empty;
+		});
+
+		it("Throws error if cached request cannot be parsed correctly", async () => {
+			const cachedRequest: TokenExchangeParameters = {
+				authority: TEST_URIS.DEFAULT_INSTANCE,
+				codeVerifier: TEST_CONFIG.TEST_VERIFIER,
+				correlationId: RANDOM_TEST_GUID,
+				scopes: [TEST_CONFIG.MSAL_CLIENT_ID],
+			};
+			const stringifiedRequest = JSON.stringify(cachedRequest);
+			defaultAuthConfig.storageInterface.setItem(TemporaryCacheKeys.REQUEST_PARAMS, stringifiedRequest.substring(0, stringifiedRequest.length / 2));
+			const codeResponse: CodeResponse = {
+				code: "This is an auth code",
+				userRequestState: RANDOM_TEST_GUID
+			};
+			await expect(Client.acquireToken(codeResponse)).to.be.rejectedWith(ClientAuthErrorMessage.tokenRequestCacheError.desc);
+			expect(defaultAuthConfig.storageInterface.getKeys()).to.be.empty;
+		});
+	});
 });
