+++ lib/msal-common/test/client/SPAClient.spec.ts
@@ -5,21 +5,17 @@ const expect = chai.expect;
 chai.use(chaiAsPromised);
 import sinon from "sinon";
 import { SPAClient } from "../../src/client/SPAClient";
-import { TEST_CONFIG, TEST_URIS, RANDOM_TEST_GUID, DEFAULT_OPENID_CONFIG_RESPONSE, TEST_TOKENS, ALTERNATE_OPENID_CONFIG_RESPONSE, TEST_DATA_CLIENT_INFO, TEST_TOKEN_LIFETIMES, TEST_HASHES } from "../utils/StringConstants";
-import { AuthenticationParameters } from "../../src/request/AuthenticationParameters";
+import { TEST_CONFIG, TEST_URIS, RANDOM_TEST_GUID, DEFAULT_OPENID_CONFIG_RESPONSE, TEST_TOKENS, ALTERNATE_OPENID_CONFIG_RESPONSE, TEST_DATA_CLIENT_INFO, TEST_TOKEN_LIFETIMES } from "../utils/StringConstants";
 import { ClientConfigurationError, ClientConfigurationErrorMessage } from "../../src/error/ClientConfigurationError";
-import { AADServerParamKeys, TemporaryCacheKeys, PersistentCacheKeys, Constants } from "../../src/utils/Constants";
-import { ServerCodeRequestParameters } from "../../src/server/ServerCodeRequestParameters";
+import { AADServerParamKeys, PersistentCacheKeys, Constants } from "../../src/utils/Constants";
 import { IdTokenClaims } from "../../src/account/IdTokenClaims";
 import { IdToken } from "../../src/account/IdToken";
 import { LogLevel } from "../../src/logger/Logger";
 import { NetworkRequestOptions } from "../../src/network/INetworkModule";
 import { Authority } from "../../src/authority/Authority";
 import { PkceCodes } from "../../src/crypto/ICrypto";
-import { TokenExchangeParameters } from "../../src/request/TokenExchangeParameters";
 import { ClientAuthError, ClientAuthErrorMessage } from "../../src/error/ClientAuthError";
 import { AuthError } from "../../src/error/AuthError";
-import { CodeResponse } from "../../src/response/CodeResponse";
 import { buildClientInfo, ClientInfo } from "../../src/account/ClientInfo";
 import { TimeUtils } from "../../src/utils/TimeUtils";
 import { AccessTokenKey } from "../../src/cache/AccessTokenKey";
@@ -32,6 +28,9 @@ import { AuthorityFactory } from "../../src/authority/AuthorityFactory";
 import { ServerError } from "../../src/error/ServerError";
 import { ClientConfiguration } from "../../src/config/ClientConfiguration";
 import { InMemoryCache } from "../../src/unifiedCache/utils/CacheTypes";
+import { AuthorizationUrlRequest } from "../../src/request/AuthorizationUrlRequest";
+import { AuthorizationCodeRequest } from "../../src/request/AuthorizationCodeRequest";
+import { AadAuthority } from "../../src/authority/AadAuthority";
 
 describe("SPAClient.ts Class Unit Tests", () => {
 
@@ -155,51 +154,45 @@ describe("SPAClient.ts Class Unit Tests", () => {
         });
 
         it("Creates a login URL with default scopes", async () => {
-            const emptyRequest: AuthenticationParameters = {};
+            const emptyRequest: AuthorizationUrlRequest = {
+				redirectUri: TEST_URIS.TEST_REDIR_URI,
+				scopes: [],
+				codeChallenge: TEST_CONFIG.TEST_CHALLENGE,
+				codeChallengeMethod: Constants.S256_CODE_CHALLENGE_METHOD
+			};
             const loginUrl = await Client.createLoginUrl(emptyRequest);
             expect(loginUrl).to.contain(Constants.DEFAULT_AUTHORITY);
             expect(loginUrl).to.contain(DEFAULT_OPENID_CONFIG_RESPONSE.body.authorization_endpoint.replace("{tenant}", "common"));
             expect(loginUrl).to.contain(`${AADServerParamKeys.SCOPE}=${Constants.OPENID_SCOPE}%20${Constants.PROFILE_SCOPE}%20${Constants.OFFLINE_ACCESS_SCOPE}`);
             expect(loginUrl).to.contain(`${AADServerParamKeys.RESPONSE_TYPE}=${Constants.CODE_RESPONSE_TYPE}`);
             expect(loginUrl).to.contain(`${AADServerParamKeys.CLIENT_ID}=${TEST_CONFIG.MSAL_CLIENT_ID}`);
-            expect(loginUrl).to.contain(`${AADServerParamKeys.REDIRECT_URI}=${encodeURIComponent(TEST_URIS.TEST_REDIR_URI)}`);
+			expect(loginUrl).to.contain(`${AADServerParamKeys.REDIRECT_URI}=${encodeURIComponent(TEST_URIS.TEST_REDIR_URI)}`);
+			expect(loginUrl).to.contain(`${AADServerParamKeys.CODE_CHALLENGE}=${encodeURIComponent(TEST_CONFIG.TEST_CHALLENGE)}`);
+			expect(loginUrl).to.contain(`${AADServerParamKeys.CODE_CHALLENGE_METHOD}=${encodeURIComponent(Constants.S256_CODE_CHALLENGE_METHOD)}`);
         });
 
         it("Creates a login URL with scopes from given token request", async () => {
             const testScope1 = "testscope1";
             const testScope2 = "testscope2";
-            const loginRequest: AuthenticationParameters = {
-                scopes: [testScope1, testScope2]
+            const loginRequest: AuthorizationUrlRequest = {
+				redirectUri: TEST_URIS.TEST_REDIR_URI,
+				scopes: [testScope1, testScope2],
+				codeChallenge: TEST_CONFIG.TEST_CHALLENGE,
+				codeChallengeMethod: Constants.S256_CODE_CHALLENGE_METHOD
             };
             const loginUrl = await Client.createLoginUrl(loginRequest);
             expect(loginUrl).to.contain(`${AADServerParamKeys.SCOPE}=${encodeURIComponent(`${testScope1} ${testScope2} ${Constants.OPENID_SCOPE} ${Constants.PROFILE_SCOPE} ${Constants.OFFLINE_ACCESS_SCOPE}`)}`);
         });
 
-        it("Updates cache entries correctly", async () => {
-            const emptyRequest: AuthenticationParameters = {};
-            await Client.createLoginUrl(emptyRequest);
-            expect(defaultAuthConfig.storageInterface.getItem(TemporaryCacheKeys.REQUEST_STATE)).to.be.deep.eq(RANDOM_TEST_GUID);
-            expect(defaultAuthConfig.storageInterface.getItem(`${TemporaryCacheKeys.NONCE_IDTOKEN}|${RANDOM_TEST_GUID}`)).to.be.eq(RANDOM_TEST_GUID);
-            expect(defaultAuthConfig.storageInterface.getItem(`${TemporaryCacheKeys.AUTHORITY}|${RANDOM_TEST_GUID}`)).to.be.eq(`${Constants.DEFAULT_AUTHORITY}/`);
-        });
-
-        it("Caches token request correctly", async () => {
-            const emptyRequest: AuthenticationParameters = {};
-            await Client.createLoginUrl(emptyRequest);
-            const cachedRequest: TokenExchangeParameters = JSON.parse(defaultAuthConfig.storageInterface.getItem(TemporaryCacheKeys.REQUEST_PARAMS));
-            expect(cachedRequest.scopes).to.be.deep.eq([TEST_CONFIG.MSAL_CLIENT_ID]);
-            expect(cachedRequest.codeVerifier).to.be.deep.eq(TEST_CONFIG.TEST_VERIFIER);
-            expect(cachedRequest.authority).to.be.deep.eq(`${Constants.DEFAULT_AUTHORITY}/`);
-            expect(cachedRequest.correlationId).to.be.deep.eq(RANDOM_TEST_GUID);
-            expect(cachedRequest.extraQueryParameters).to.be.undefined;
-            expect(cachedRequest.resource).to.be.undefined;
-        });
-
         it("Uses authority if given in request", async () => {
             sinon.restore();
             sinon.stub(Authority.prototype, <any>"discoverEndpoints").resolves(ALTERNATE_OPENID_CONFIG_RESPONSE);
-            const loginRequest: AuthenticationParameters = {
-                authority: `${TEST_URIS.ALTERNATE_INSTANCE}/common`
+            const loginRequest: AuthorizationUrlRequest = {
+				redirectUri: TEST_URIS.TEST_REDIR_URI,
+                scopes: [],
+				authority: `${TEST_URIS.ALTERNATE_INSTANCE}/common`,
+				codeChallenge: TEST_CONFIG.TEST_CHALLENGE,
+				codeChallengeMethod: Constants.S256_CODE_CHALLENGE_METHOD
             };
             const loginUrl = await Client.createLoginUrl(loginRequest);
             expect(loginUrl).to.contain(TEST_URIS.ALTERNATE_INSTANCE);
@@ -214,64 +207,13 @@ describe("SPAClient.ts Class Unit Tests", () => {
             sinon.restore();
             const exceptionString = "Could not make a network request.";
             sinon.stub(Authority.prototype, "resolveEndpointsAsync").throwsException(exceptionString);
-            const emptyRequest: AuthenticationParameters = {};
+            const emptyRequest: AuthorizationUrlRequest = {
+				redirectUri: TEST_URIS.TEST_REDIR_URI,
+				scopes: [],
+				codeChallenge: TEST_CONFIG.TEST_CHALLENGE,
+				codeChallengeMethod: Constants.S256_CODE_CHALLENGE_METHOD
+			};
             await expect(Client.createLoginUrl(emptyRequest)).to.be.rejectedWith(`${ClientAuthErrorMessage.endpointResolutionError.desc} Detail: ${exceptionString}`);
-        });
-
-        it("Cleans cache before error is thrown", async () => {
-            const guidCreationErr = "GUID can't be created.";
-            const emptyRequest: AuthenticationParameters = {};
-            defaultAuthConfig.cryptoInterface.createNewGuid = (): string => {
-                throw AuthError.createUnexpectedError(guidCreationErr);
-            };
-            Client = new SPAClient(defaultAuthConfig);
-            await expect(Client.createLoginUrl(emptyRequest)).to.be.rejectedWith(guidCreationErr);
-            expect(defaultAuthConfig.storageInterface.getKeys()).to.be.empty;
-        });
-
-        it("Uses adal token from cache if it is present.", async () => {
-            const idTokenClaims: IdTokenClaims = {
-                "iss": "https://sts.windows.net/fa15d692-e9c7-4460-a743-29f2956fd429/",
-                "exp": "1536279024",
-                "name": "abeli",
-                "nonce": "123523",
-                "oid": "05833b6b-aa1d-42d4-9ec0-1b2bb9194438",
-                "sub": "5_J9rSss8-jvt_Icu6ueRNL8xXb8LF4Fsg_KooC2RJQ",
-                "tid": "fa15d692-e9c7-4460-a743-29f2956fd429",
-                "ver": "1.0"
-            };
-            sinon.stub(IdToken, "extractIdToken").returns(idTokenClaims);
-            defaultAuthConfig.storageInterface.setItem(PersistentCacheKeys.ADAL_ID_TOKEN, TEST_TOKENS.IDTOKEN_V1);
-            const testToken = new IdToken(TEST_TOKENS.IDTOKEN_V1, defaultAuthConfig.cryptoInterface);
-            const queryParamSpy = sinon.spy(ServerCodeRequestParameters.prototype, "populateQueryParams");
-            Client = new SPAClient(defaultAuthConfig);
-            const emptyRequest: AuthenticationParameters = {};
-            await Client.createLoginUrl(emptyRequest);
-            expect(queryParamSpy.calledWith(testToken)).to.be.true;
-        });
-
-        it("Does not use adal token from cache if it is present and SSO params have been given.", async () => {
-            const idTokenClaims: IdTokenClaims = {
-                "iss": "https://sts.windows.net/fa15d692-e9c7-4460-a743-29f2956fd429/",
-                "exp": "1536279024",
-                "name": "abeli",
-                "nonce": "123523",
-                "oid": "05833b6b-aa1d-42d4-9ec0-1b2bb9194438",
-                "sub": "5_J9rSss8-jvt_Icu6ueRNL8xXb8LF4Fsg_KooC2RJQ",
-                "tid": "fa15d692-e9c7-4460-a743-29f2956fd429",
-                "ver": "1.0"
-            };
-            sinon.stub(IdToken, "extractIdToken").returns(idTokenClaims);
-            defaultAuthConfig.storageInterface.setItem(PersistentCacheKeys.ADAL_ID_TOKEN, TEST_TOKENS.IDTOKEN_V1);
-            const testToken = new IdToken(TEST_TOKENS.IDTOKEN_V1, defaultAuthConfig.cryptoInterface);
-            const queryParamSpy = sinon.spy(ServerCodeRequestParameters.prototype, "populateQueryParams");
-            Client = new SPAClient(defaultAuthConfig);
-            const loginRequest: AuthenticationParameters = {
-                loginHint: "AbeLi@microsoft.com"
-            };
-            await Client.createLoginUrl(loginRequest);
-            expect(queryParamSpy.calledWith(testToken)).to.be.false;
-            expect(queryParamSpy.calledWith(null)).to.be.true;
         });
     });
 
@@ -290,8 +232,11 @@ describe("SPAClient.ts Class Unit Tests", () => {
         it("Creates a acquire token URL with scopes from given token request", async () => {
             const testScope1 = "testscope1";
             const testScope2 = "testscope2";
-            const tokenRequest: AuthenticationParameters = {
-                scopes: [testScope1, testScope2]
+            const tokenRequest: AuthorizationUrlRequest = {
+				redirectUri: TEST_URIS.TEST_REDIR_URI,
+				scopes: [testScope1, testScope2],
+				codeChallenge: TEST_CONFIG.TEST_CHALLENGE,
+				codeChallengeMethod: Constants.S256_CODE_CHALLENGE_METHOD
             };
             const acquireTokenUrl = await Client.createAcquireTokenUrl(tokenRequest);
             expect(acquireTokenUrl).to.contain(Constants.DEFAULT_AUTHORITY);
@@ -303,8 +248,11 @@ describe("SPAClient.ts Class Unit Tests", () => {
         });
 
         it("Creates a acquire token URL with replaced client id scope", async () => {
-            const tokenRequest: AuthenticationParameters = {
-                scopes: [TEST_CONFIG.MSAL_CLIENT_ID]
+            const tokenRequest: AuthorizationUrlRequest = {
+				redirectUri: TEST_URIS.TEST_REDIR_URI,
+				scopes: [TEST_CONFIG.MSAL_CLIENT_ID],
+				codeChallenge: TEST_CONFIG.TEST_CHALLENGE,
+				codeChallengeMethod: Constants.S256_CODE_CHALLENGE_METHOD
             };
             const acquireTokenUrl = await Client.createAcquireTokenUrl(tokenRequest);
             expect(acquireTokenUrl).to.contain(Constants.DEFAULT_AUTHORITY);
@@ -316,49 +264,34 @@ describe("SPAClient.ts Class Unit Tests", () => {
         });
 
         it("Throws error if no scopes are passed to createAcquireTokenUrl", async () => {
-            const emptyRequest: AuthenticationParameters = {};
+            const emptyRequest: AuthorizationUrlRequest = {
+				redirectUri: TEST_URIS.TEST_REDIR_URI,
+				scopes: null,
+				codeChallenge: TEST_CONFIG.TEST_CHALLENGE,
+				codeChallengeMethod: Constants.S256_CODE_CHALLENGE_METHOD
+			};
             await expect(Client.createAcquireTokenUrl(emptyRequest)).to.be.rejectedWith(ClientConfigurationErrorMessage.emptyScopesError.desc);
         });
 
         it("Throws error if empty scopes are passed to createAcquireTokenUrl", async () => {
-            const emptyRequest: AuthenticationParameters = {
-                scopes: []
-            };
+            const emptyRequest: AuthorizationUrlRequest = {
+				redirectUri: TEST_URIS.TEST_REDIR_URI,
+				scopes: [],
+				codeChallenge: TEST_CONFIG.TEST_CHALLENGE,
+				codeChallengeMethod: Constants.S256_CODE_CHALLENGE_METHOD
+			};
             await expect(Client.createAcquireTokenUrl(emptyRequest)).to.be.rejectedWith(ClientConfigurationErrorMessage.emptyScopesError.desc);
-        });
-
-        it("Updates cache entries correctly", async () => {
-            const testScope = "testscope";
-            const tokenRequest: AuthenticationParameters = {
-                scopes: [testScope]
-            };
-            await Client.createAcquireTokenUrl(tokenRequest);
-            expect(defaultAuthConfig.storageInterface.getItem(TemporaryCacheKeys.REQUEST_STATE)).to.be.deep.eq(RANDOM_TEST_GUID);
-            expect(defaultAuthConfig.storageInterface.getItem(`${TemporaryCacheKeys.NONCE_IDTOKEN}|${RANDOM_TEST_GUID}`)).to.be.eq(RANDOM_TEST_GUID);
-            expect(defaultAuthConfig.storageInterface.getItem(`${TemporaryCacheKeys.AUTHORITY}|${RANDOM_TEST_GUID}`)).to.be.eq(`${Constants.DEFAULT_AUTHORITY}/`);
-        });
-
-        it("Caches token request correctly", async () => {
-            const testScope = "testscope";
-            const tokenRequest: AuthenticationParameters = {
-                scopes: [testScope]
-            };
-            await Client.createAcquireTokenUrl(tokenRequest);
-            const cachedRequest: TokenExchangeParameters = JSON.parse(defaultAuthConfig.storageInterface.getItem(TemporaryCacheKeys.REQUEST_PARAMS));
-            expect(cachedRequest.scopes).to.be.deep.eq([testScope]);
-            expect(cachedRequest.codeVerifier).to.be.deep.eq(TEST_CONFIG.TEST_VERIFIER);
-            expect(cachedRequest.authority).to.be.deep.eq(`${Constants.DEFAULT_AUTHORITY}/`);
-            expect(cachedRequest.correlationId).to.be.deep.eq(RANDOM_TEST_GUID);
-            expect(cachedRequest.extraQueryParameters).to.be.undefined;
-            expect(cachedRequest.resource).to.be.undefined;
         });
 
         it("Uses authority if given in request", async () => {
             sinon.restore();
             sinon.stub(Authority.prototype, <any>"discoverEndpoints").resolves(ALTERNATE_OPENID_CONFIG_RESPONSE);
-            const tokenRequest: AuthenticationParameters = {
-                authority: `${TEST_URIS.ALTERNATE_INSTANCE}/common`,
-                scopes: [TEST_CONFIG.MSAL_CLIENT_ID]
+            const tokenRequest: AuthorizationUrlRequest = {
+				authority: `${TEST_URIS.ALTERNATE_INSTANCE}/common`,
+				redirectUri: TEST_URIS.TEST_REDIR_URI,
+				scopes: [TEST_CONFIG.MSAL_CLIENT_ID],
+				codeChallenge: TEST_CONFIG.TEST_CHALLENGE,
+				codeChallengeMethod: Constants.S256_CODE_CHALLENGE_METHOD
             };
             const acquireTokenUrl = await Client.createAcquireTokenUrl(tokenRequest);
             expect(acquireTokenUrl).to.contain(TEST_URIS.ALTERNATE_INSTANCE);
@@ -373,16 +306,22 @@ describe("SPAClient.ts Class Unit Tests", () => {
             sinon.restore();
             const exceptionString = "Could not make a network request.";
             sinon.stub(Authority.prototype, "resolveEndpointsAsync").throwsException(exceptionString);
-            const tokenRequest: AuthenticationParameters = {
-                scopes: [TEST_CONFIG.MSAL_CLIENT_ID]
+            const tokenRequest: AuthorizationUrlRequest = {
+				redirectUri: TEST_URIS.TEST_REDIR_URI,
+				scopes: [TEST_CONFIG.MSAL_CLIENT_ID],
+				codeChallenge: TEST_CONFIG.TEST_CHALLENGE,
+				codeChallengeMethod: Constants.S256_CODE_CHALLENGE_METHOD
             };
             await expect(Client.createAcquireTokenUrl(tokenRequest)).to.be.rejectedWith(`${ClientAuthErrorMessage.endpointResolutionError.desc} Detail: ${exceptionString}`);
         });
 
         it("Cleans cache before error is thrown", async () => {
             const guidCreationErr = "GUID can't be created.";
-            const tokenRequest: AuthenticationParameters = {
-                scopes: [TEST_CONFIG.MSAL_CLIENT_ID]
+            const tokenRequest: AuthorizationUrlRequest = {
+				redirectUri: TEST_URIS.TEST_REDIR_URI,
+				scopes: [TEST_CONFIG.MSAL_CLIENT_ID],
+				codeChallenge: TEST_CONFIG.TEST_CHALLENGE,
+				codeChallengeMethod: Constants.S256_CODE_CHALLENGE_METHOD
             };
             defaultAuthConfig.cryptoInterface.createNewGuid = (): string => {
                 throw AuthError.createUnexpectedError(guidCreationErr);
@@ -390,54 +329,6 @@ describe("SPAClient.ts Class Unit Tests", () => {
             Client = new SPAClient(defaultAuthConfig);
             await expect(Client.createAcquireTokenUrl(tokenRequest)).to.be.rejectedWith(guidCreationErr);
             expect(defaultAuthConfig.storageInterface.getKeys()).to.be.empty;
-        });
-
-        it("Uses adal token from cache if it is present.", async () => {
-            const idTokenClaims: IdTokenClaims = {
-                "iss": "https://sts.windows.net/fa15d692-e9c7-4460-a743-29f2956fd429/",
-                "exp": "1536279024",
-                "name": "abeli",
-                "nonce": "123523",
-                "oid": "05833b6b-aa1d-42d4-9ec0-1b2bb9194438",
-                "sub": "5_J9rSss8-jvt_Icu6ueRNL8xXb8LF4Fsg_KooC2RJQ",
-                "tid": "fa15d692-e9c7-4460-a743-29f2956fd429",
-                "ver": "1.0"
-            };
-            sinon.stub(IdToken, "extractIdToken").returns(idTokenClaims);
-            defaultAuthConfig.storageInterface.setItem(PersistentCacheKeys.ADAL_ID_TOKEN, TEST_TOKENS.IDTOKEN_V1);
-            const testToken = new IdToken(TEST_TOKENS.IDTOKEN_V1, defaultAuthConfig.cryptoInterface);
-            const queryParamSpy = sinon.spy(ServerCodeRequestParameters.prototype, "populateQueryParams");
-            Client = new SPAClient(defaultAuthConfig);
-            const tokenRequest: AuthenticationParameters = {
-                scopes: [TEST_CONFIG.MSAL_CLIENT_ID]
-            };
-            await Client.createAcquireTokenUrl(tokenRequest);
-            expect(queryParamSpy.calledWith(testToken)).to.be.true;
-        });
-
-        it("Does not use adal token from cache if it is present and SSO params have been given.", async () => {
-            const idTokenClaims: IdTokenClaims = {
-                "iss": "https://sts.windows.net/fa15d692-e9c7-4460-a743-29f2956fd429/",
-                "exp": "1536279024",
-                "name": "abeli",
-                "nonce": "123523",
-                "oid": "05833b6b-aa1d-42d4-9ec0-1b2bb9194438",
-                "sub": "5_J9rSss8-jvt_Icu6ueRNL8xXb8LF4Fsg_KooC2RJQ",
-                "tid": "fa15d692-e9c7-4460-a743-29f2956fd429",
-                "ver": "1.0"
-            };
-            sinon.stub(IdToken, "extractIdToken").returns(idTokenClaims);
-            defaultAuthConfig.storageInterface.setItem(PersistentCacheKeys.ADAL_ID_TOKEN, TEST_TOKENS.IDTOKEN_V1);
-            const testToken = new IdToken(TEST_TOKENS.IDTOKEN_V1, defaultAuthConfig.cryptoInterface);
-            const queryParamSpy = sinon.spy(ServerCodeRequestParameters.prototype, "populateQueryParams");
-            Client = new SPAClient(defaultAuthConfig);
-            const tokenRequest: AuthenticationParameters = {
-                scopes: [TEST_CONFIG.MSAL_CLIENT_ID],
-                loginHint: "AbeLi@microsoft.com"
-            };
-            await Client.createAcquireTokenUrl(tokenRequest);
-            expect(queryParamSpy.calledWith(testToken)).to.be.false;
-            expect(queryParamSpy.calledWith(null)).to.be.true;
         });
     });
 
@@ -456,70 +347,41 @@ describe("SPAClient.ts Class Unit Tests", () => {
 
             describe("Error Cases", () => {
 
-                it("Throws error if null code response is passed", async () => {
-                    await expect(Client.acquireToken(null)).to.be.rejectedWith(ClientAuthErrorMessage.tokenRequestCannotBeMade.desc);
+                it("Throws error if null code request is passed", async () => {
+                    await expect(Client.acquireToken(null, "", "")).to.be.rejectedWith(ClientAuthErrorMessage.tokenRequestCannotBeMade.desc);
                     expect(defaultAuthConfig.storageInterface.getKeys()).to.be.empty;
                 });
 
                 it("Throws error if code response does not contain PKCE code", async () => {
-                    const codeResponse: CodeResponse = {
-                        code: null,
-                        userRequestState: RANDOM_TEST_GUID
+                    const codeRequest: AuthorizationCodeRequest = {
+						redirectUri: TEST_URIS.TEST_REDIR_URI,
+						scopes: ["scope"],
+                        code: null
                     };
-                    await expect(Client.acquireToken(codeResponse)).to.be.rejectedWith(ClientAuthErrorMessage.tokenRequestCannotBeMade.desc);
-                    expect(defaultAuthConfig.storageInterface.getKeys()).to.be.empty;
-                });
-
-                it("Throws error if request cannot be retrieved from cache", async () => {
-                    const codeResponse: CodeResponse = {
-                        code: "This is an auth code",
-                        userRequestState: RANDOM_TEST_GUID
-                    };
-                    await expect(Client.acquireToken(codeResponse)).to.be.rejectedWith(ClientAuthErrorMessage.tokenRequestCacheError.desc);
-                    expect(defaultAuthConfig.storageInterface.getKeys()).to.be.empty;
-                });
-
-                it("Throws error if cached request cannot be parsed correctly", async () => {
-                    const cachedRequest: TokenExchangeParameters = {
-                        authority: TEST_URIS.DEFAULT_INSTANCE,
-                        codeVerifier: TEST_CONFIG.TEST_VERIFIER,
-                        correlationId: RANDOM_TEST_GUID,
-                        scopes: [TEST_CONFIG.MSAL_CLIENT_ID],
-                    };
-                    const stringifiedRequest = JSON.stringify(cachedRequest);
-                    defaultAuthConfig.storageInterface.setItem(TemporaryCacheKeys.REQUEST_PARAMS, stringifiedRequest.substring(0, stringifiedRequest.length / 2));
-                    const codeResponse: CodeResponse = {
-                        code: "This is an auth code",
-                        userRequestState: RANDOM_TEST_GUID
-                    };
-                    await expect(Client.acquireToken(codeResponse)).to.be.rejectedWith(ClientAuthErrorMessage.tokenRequestCacheError.desc);
+                    await expect(Client.acquireToken(codeRequest, "", "")).to.be.rejectedWith(ClientAuthErrorMessage.tokenRequestCannotBeMade.desc);
                     expect(defaultAuthConfig.storageInterface.getKeys()).to.be.empty;
                 });
 
                 it("Throws error if authority endpoint resolution fails", async () => {
-                    const codeResponse = {
-                        code: "This is an auth code",
-                        userRequestState: RANDOM_TEST_GUID
+					const codeRequest: AuthorizationCodeRequest = {
+						authority: Constants.DEFAULT_AUTHORITY,
+                        codeVerifier: TEST_CONFIG.TEST_VERIFIER,
+                        correlationId: RANDOM_TEST_GUID,
+                        scopes: [TEST_CONFIG.MSAL_CLIENT_ID],
+						redirectUri: TEST_URIS.TEST_REDIR_URI,
+                        code: "This is an auth code"
                     };
                     const exceptionString = "Could not make a network request.";
                     sinon.stub(Authority.prototype, "resolveEndpointsAsync").throwsException(exceptionString);
 
-                    const cachedRequest: TokenExchangeParameters = {
-                        authority: Constants.DEFAULT_AUTHORITY,
-                        codeVerifier: TEST_CONFIG.TEST_VERIFIER,
-                        correlationId: RANDOM_TEST_GUID,
-                        scopes: [TEST_CONFIG.MSAL_CLIENT_ID],
-                    };
-                    const stringifiedRequest = JSON.stringify(cachedRequest);
-                    defaultAuthConfig.storageInterface.setItem(TemporaryCacheKeys.REQUEST_PARAMS, stringifiedRequest);
-                    await expect(Client.acquireToken(codeResponse)).to.be.rejectedWith(`${ClientAuthErrorMessage.endpointResolutionError.desc} Detail: ${exceptionString}`);
+                    await expect(Client.acquireToken(codeRequest, RANDOM_TEST_GUID, RANDOM_TEST_GUID)).to.be.rejectedWith(`${ClientAuthErrorMessage.endpointResolutionError.desc} Detail: ${exceptionString}`);
                 });
             });
 
             describe("Success Cases", () => {
 
-                let codeResponse: CodeResponse;
-                let testState: string;
+				let testState: string;
+				let codeRequest: AuthorizationCodeRequest;
                 beforeEach(() => {
                     // Set up required objects and mocked return values
                     defaultAuthConfig.networkInterface.sendPostRequestAsync = (url: string, options?: NetworkRequestOptions): any => {
@@ -556,9 +418,13 @@ describe("SPAClient.ts Class Unit Tests", () => {
                     Client = new SPAClient(defaultAuthConfig);
 
                     testState = "{stateObject}";
-                    codeResponse = {
-                        code: "This is an auth code",
-                        userRequestState: `${RANDOM_TEST_GUID}|${testState}`
+                    codeRequest = {
+						authority: Constants.DEFAULT_AUTHORITY,
+                        codeVerifier: TEST_CONFIG.TEST_VERIFIER,
+                        correlationId: RANDOM_TEST_GUID,
+                        scopes: [TEST_CONFIG.MSAL_CLIENT_ID],
+						redirectUri: TEST_URIS.TEST_REDIR_URI,
+                        code: "This is an auth code"
                     };
                 });
 
@@ -579,7 +445,6 @@ describe("SPAClient.ts Class Unit Tests", () => {
                     sinon.stub(Authority.prototype, <any>"discoverEndpoints").resolves(DEFAULT_OPENID_CONFIG_RESPONSE);
 
                     // Set up cache
-                    defaultAuthConfig.storageInterface.setItem(`${TemporaryCacheKeys.NONCE_IDTOKEN}|${codeResponse.userRequestState}`, "123523");
                     defaultAuthConfig.storageInterface.setItem(PersistentCacheKeys.CLIENT_INFO, TEST_DATA_CLIENT_INFO.TEST_RAW_CLIENT_INFO);
 
                     // Build Test account
@@ -587,18 +452,8 @@ describe("SPAClient.ts Class Unit Tests", () => {
                     const clientInfo = buildClientInfo(TEST_DATA_CLIENT_INFO.TEST_RAW_CLIENT_INFO, defaultAuthConfig.cryptoInterface);
                     const testAccount = Account.createAccount(idToken, clientInfo, defaultAuthConfig.cryptoInterface);
 
-                    const cachedRequest: TokenExchangeParameters = {
-                        authority: Constants.DEFAULT_AUTHORITY,
-                        codeVerifier: TEST_CONFIG.TEST_VERIFIER,
-                        correlationId: RANDOM_TEST_GUID,
-                        scopes: [TEST_CONFIG.MSAL_CLIENT_ID],
-                    };
-                    const stringifiedRequest = JSON.stringify(cachedRequest);
-                    defaultAuthConfig.storageInterface.setItem(TemporaryCacheKeys.REQUEST_PARAMS, stringifiedRequest);
-                    defaultAuthConfig.storageInterface.setItem(`${TemporaryCacheKeys.AUTHORITY}${Constants.RESOURCE_DELIM}${RANDOM_TEST_GUID}`, `${TEST_URIS.DEFAULT_INSTANCE}/common/`);
-
                     // Perform test
-                    const tokenResponse: TokenResponse = await Client.acquireToken(codeResponse);
+                    const tokenResponse: TokenResponse = await Client.acquireToken(codeRequest, testState, idTokenClaims.nonce);
                     expect(tokenResponse.uniqueId).to.be.deep.eq(idTokenClaims.oid);
                     expect(tokenResponse.tenantId).to.be.deep.eq(idTokenClaims.tid);
                     expect(tokenResponse.tokenType).to.be.deep.eq(TEST_CONFIG.TOKEN_TYPE_BEARER);
@@ -611,41 +466,6 @@ describe("SPAClient.ts Class Unit Tests", () => {
                     expect(tokenResponse.scopes).to.be.deep.eq(TEST_CONFIG.DEFAULT_SCOPES);
                     expect(tokenResponse.userRequestState).to.be.deep.eq(testState);
                     expect(Account.compareAccounts(Client.getAccount(), testAccount)).to.be.true;
-                });
-
-                it("Uses authority from cache if not present in cached request", async () => {
-                    // Set up stubs
-                    const idTokenClaims = {
-                        "ver": "2.0",
-                        "iss": `${TEST_URIS.ALTERNATE_INSTANCE}9188040d-6c67-4c5b-b112-36a304b66dad/v2.0`,
-                        "sub": "AAAAAAAAAAAAAAAAAAAAAIkzqFVrSaSaFHy782bbtaQ",
-                        "exp": "1536361411",
-                        "name": "Abe Lincoln",
-                        "preferred_username": "AbeLi@microsoft.com",
-                        "oid": "00000000-0000-0000-66f3-3332eca7ea81",
-                        "tid": "3338040d-6c67-4c5b-b112-36a304b66dad",
-                        "nonce": "123523",
-                    };
-                    sinon.stub(IdToken, "extractIdToken").returns(idTokenClaims);
-                    sinon.stub(Authority.prototype, <any>"discoverEndpoints").resolves(ALTERNATE_OPENID_CONFIG_RESPONSE);
-                    const authoritySpy = sinon.spy(AuthorityFactory, "createInstance");
-
-                    // Set up cache
-                    defaultAuthConfig.storageInterface.setItem(`${TemporaryCacheKeys.NONCE_IDTOKEN}|${codeResponse.userRequestState}`, "123523");
-                    defaultAuthConfig.storageInterface.setItem(PersistentCacheKeys.CLIENT_INFO, TEST_DATA_CLIENT_INFO.TEST_RAW_CLIENT_INFO);
-
-                    const cachedRequest: TokenExchangeParameters = {
-                        codeVerifier: TEST_CONFIG.TEST_VERIFIER,
-                        correlationId: RANDOM_TEST_GUID,
-                        scopes: [TEST_CONFIG.MSAL_CLIENT_ID],
-                    };
-                    const stringifiedRequest = JSON.stringify(cachedRequest);
-                    defaultAuthConfig.storageInterface.setItem(TemporaryCacheKeys.REQUEST_PARAMS, stringifiedRequest);
-                    defaultAuthConfig.storageInterface.setItem(`${TemporaryCacheKeys.AUTHORITY}${Constants.RESOURCE_DELIM}${codeResponse.userRequestState}`, `${TEST_URIS.ALTERNATE_INSTANCE}/common/`);
-
-                    // Perform test
-                    await Client.acquireToken(codeResponse);
-                    expect(authoritySpy.calledOnceWith(`${TEST_URIS.ALTERNATE_INSTANCE}/common/`, defaultAuthConfig.networkInterface)).to.be.true;
                 });
             });
         });
@@ -665,44 +485,60 @@ describe("SPAClient.ts Class Unit Tests", () => {
             describe("Error cases", () => {
 
                 it("Throws error if request object is null or undefined", async () => {
-                    await expect(authClient.getValidToken(null)).to.be.rejectedWith(ClientConfigurationErrorMessage.tokenRequestEmptyError.desc);
-                    await expect(authClient.getValidToken(undefined)).to.be.rejectedWith(ClientConfigurationErrorMessage.tokenRequestEmptyError.desc);
+                    await expect(authClient.getValidToken(null, null, false)).to.be.rejectedWith(ClientConfigurationErrorMessage.tokenRequestEmptyError.desc);
+                    await expect(authClient.getValidToken(null, null, false)).to.be.rejectedWith(ClientConfigurationErrorMessage.tokenRequestEmptyError.desc);
                 });
 
                 it("Throws error if scopes are not included in request object", async () => {
-                    await expect(authClient.getValidToken({})).to.be.rejectedWith(ClientConfigurationErrorMessage.emptyScopesError.desc);
+                    await expect(authClient.getValidToken({
+						redirectUri: TEST_URIS.TEST_REDIR_URI,
+						scopes: null,
+						codeChallenge: TEST_CONFIG.TEST_CHALLENGE,
+						codeChallengeMethod: Constants.S256_CODE_CHALLENGE_METHOD
+					}, null, false)).to.be.rejectedWith(ClientConfigurationErrorMessage.emptyScopesError.desc);
                 });
 
                 it("Throws error if scopes are empty in request object", async () => {
-                    const tokenRequest: TokenRenewParameters = {
-                        scopes: []
+                    const tokenRequest: AuthorizationUrlRequest = {
+						redirectUri: TEST_URIS.TEST_REDIR_URI,
+						scopes: [],
+						codeChallenge: TEST_CONFIG.TEST_CHALLENGE,
+						codeChallengeMethod: Constants.S256_CODE_CHALLENGE_METHOD
                     };
-                    await expect(authClient.getValidToken(tokenRequest)).to.be.rejectedWith(ClientConfigurationErrorMessage.emptyScopesError.desc);
+                    await expect(authClient.getValidToken(tokenRequest, null, false)).to.be.rejectedWith(ClientConfigurationErrorMessage.emptyScopesError.desc);
                 });
 
                 it("Throws error if login hasn't been completed and client id is passed as scope", async () => {
-                    const tokenRequest: TokenRenewParameters = {
-                        scopes: [TEST_CONFIG.MSAL_CLIENT_ID]
+                    const tokenRequest: AuthorizationUrlRequest = {
+						redirectUri: TEST_URIS.TEST_REDIR_URI,
+						scopes: [TEST_CONFIG.MSAL_CLIENT_ID],
+						codeChallenge: TEST_CONFIG.TEST_CHALLENGE,
+						codeChallengeMethod: Constants.S256_CODE_CHALLENGE_METHOD
                     };
-                    await expect(authClient.getValidToken(tokenRequest)).to.be.rejectedWith(ClientAuthErrorMessage.userLoginRequiredError.desc);
+                    await expect(authClient.getValidToken(tokenRequest, null, false)).to.be.rejectedWith(ClientAuthErrorMessage.userLoginRequiredError.desc);
                 });
 
                 it("Throws error if endpoint discovery could not be completed", async () => {
                     const exceptionString = "Could not make a network request.";
                     sinon.stub(Authority.prototype, "resolveEndpointsAsync").throwsException(exceptionString);
-
-                    const tokenRequest: TokenRenewParameters = {
-                        scopes: ["scope1"]
+					const tokenRequest: AuthorizationUrlRequest = {
+						redirectUri: TEST_URIS.TEST_REDIR_URI,
+						scopes: ["scope1"],
+						codeChallenge: TEST_CONFIG.TEST_CHALLENGE,
+						codeChallengeMethod: Constants.S256_CODE_CHALLENGE_METHOD
                     };
-                    await expect(authClient.getValidToken(tokenRequest)).to.be.rejectedWith(`${ClientAuthErrorMessage.endpointResolutionError.desc} Detail: ${exceptionString}`);
+                    await expect(authClient.getValidToken(tokenRequest, null, false)).to.be.rejectedWith(`${ClientAuthErrorMessage.endpointResolutionError.desc} Detail: ${exceptionString}`);
                 });
 
                 it("Throws error if it does not find token in empty cache", async () => {
                     sinon.stub(Authority.prototype, <any>"discoverEndpoints").resolves(DEFAULT_OPENID_CONFIG_RESPONSE);
-                    const tokenRequest: TokenRenewParameters = {
-                        scopes: ["scope1"]
+                    const tokenRequest: AuthorizationUrlRequest = {
+						redirectUri: TEST_URIS.TEST_REDIR_URI,
+						scopes: ["scope1"],
+						codeChallenge: TEST_CONFIG.TEST_CHALLENGE,
+						codeChallengeMethod: Constants.S256_CODE_CHALLENGE_METHOD
                     };
-                    await expect(authClient.getValidToken(tokenRequest)).to.be.rejectedWith(ClientAuthErrorMessage.noTokensFoundError.desc);
+                    await expect(authClient.getValidToken(tokenRequest, null, false)).to.be.rejectedWith(ClientAuthErrorMessage.noTokensFoundError.desc);
                 });
 
                 it("Throws error if it does not find token in non-empty cache", async () => {
@@ -712,8 +548,7 @@ describe("SPAClient.ts Class Unit Tests", () => {
                         clientId: TEST_CONFIG.MSAL_CLIENT_ID,
                         scopes: testScope1,
                         authority: `${Constants.DEFAULT_AUTHORITY}/`,
-                        homeAccountIdentifier: TEST_DATA_CLIENT_INFO.TEST_HOME_ACCOUNT_ID,
-                        resource: "Resource1"
+                        homeAccountIdentifier: TEST_DATA_CLIENT_INFO.TEST_HOME_ACCOUNT_ID
                     };
                     const atValue: AccessTokenValue = {
                         accessToken: TEST_TOKENS.ACCESS_TOKEN,
@@ -724,11 +559,14 @@ describe("SPAClient.ts Class Unit Tests", () => {
                         extExpiresOnSec: `${TimeUtils.nowSeconds() + TEST_TOKEN_LIFETIMES.DEFAULT_EXPIRES_IN + TEST_TOKEN_LIFETIMES.DEFAULT_EXPIRES_IN}`
                     };
                     defaultAuthConfig.storageInterface.setItem(JSON.stringify(accessTokenKey1), JSON.stringify(atValue));
-                    sinon.stub(Authority.prototype, <any>"discoverEndpoints").resolves(DEFAULT_OPENID_CONFIG_RESPONSE);
-                    const tokenRequest: TokenRenewParameters = {
-                        scopes: [testScope2]
+					sinon.stub(Authority.prototype, <any>"discoverEndpoints").resolves(DEFAULT_OPENID_CONFIG_RESPONSE);
+					const tokenRequest: AuthorizationUrlRequest = {
+						redirectUri: TEST_URIS.TEST_REDIR_URI,
+						scopes: [testScope2],
+						codeChallenge: TEST_CONFIG.TEST_CHALLENGE,
+						codeChallengeMethod: Constants.S256_CODE_CHALLENGE_METHOD
                     };
-                    await expect(authClient.getValidToken(tokenRequest)).to.be.rejectedWith(ClientAuthErrorMessage.noTokensFoundError.desc);
+                    await expect(authClient.getValidToken(tokenRequest, null, false)).to.be.rejectedWith(ClientAuthErrorMessage.noTokensFoundError.desc);
                 });
 
                 it("Throws error if it finds too many tokens in cache for the same scope and client id but no authority, resource or account is given", async () => {
@@ -737,15 +575,13 @@ describe("SPAClient.ts Class Unit Tests", () => {
                         clientId: TEST_CONFIG.MSAL_CLIENT_ID,
                         scopes: testScope,
                         authority: `${Constants.DEFAULT_AUTHORITY}/`,
-                        homeAccountIdentifier: TEST_DATA_CLIENT_INFO.TEST_HOME_ACCOUNT_ID,
-                        resource: "Resource1"
+                        homeAccountIdentifier: TEST_DATA_CLIENT_INFO.TEST_HOME_ACCOUNT_ID
                     };
                     const accessTokenKey2: AccessTokenKey = {
                         clientId: TEST_CONFIG.MSAL_CLIENT_ID,
                         scopes: testScope,
                         authority: `${Constants.DEFAULT_AUTHORITY}/`,
-                        homeAccountIdentifier: TEST_DATA_CLIENT_INFO.TEST_HOME_ACCOUNT_ID,
-                        resource: "Resource2"
+                        homeAccountIdentifier: "homeAccountId2"
                     };
                     const atValue: AccessTokenValue = {
                         accessToken: TEST_TOKENS.ACCESS_TOKEN,
@@ -758,10 +594,13 @@ describe("SPAClient.ts Class Unit Tests", () => {
                     defaultAuthConfig.storageInterface.setItem(JSON.stringify(accessTokenKey1), JSON.stringify(atValue));
                     defaultAuthConfig.storageInterface.setItem(JSON.stringify(accessTokenKey2), JSON.stringify(atValue));
                     sinon.stub(Authority.prototype, <any>"discoverEndpoints").resolves(DEFAULT_OPENID_CONFIG_RESPONSE);
-                    const tokenRequest: TokenRenewParameters = {
-                        scopes: [testScope]
+                    const tokenRequest: AuthorizationUrlRequest = {
+						redirectUri: TEST_URIS.TEST_REDIR_URI,
+						scopes: [testScope],
+						codeChallenge: TEST_CONFIG.TEST_CHALLENGE,
+						codeChallengeMethod: Constants.S256_CODE_CHALLENGE_METHOD
                     };
-                    await expect(authClient.getValidToken(tokenRequest)).to.be.rejectedWith(ClientAuthErrorMessage.multipleMatchingTokens.desc);
+                    await expect(authClient.getValidToken(tokenRequest, null, false)).to.be.rejectedWith(ClientAuthErrorMessage.multipleMatchingTokens.desc);
                 });
             });
 
@@ -773,8 +612,7 @@ describe("SPAClient.ts Class Unit Tests", () => {
                         clientId: TEST_CONFIG.MSAL_CLIENT_ID,
                         scopes: `${testScope1} ${Constants.OFFLINE_ACCESS_SCOPE}`,
                         authority: `${Constants.DEFAULT_AUTHORITY}/`,
-                        homeAccountIdentifier: TEST_DATA_CLIENT_INFO.TEST_HOME_ACCOUNT_ID,
-                        resource: "Resource1"
+                        homeAccountIdentifier: TEST_DATA_CLIENT_INFO.TEST_HOME_ACCOUNT_ID
                     };
                     const atValue: AccessTokenValue = {
                         accessToken: TEST_TOKENS.ACCESS_TOKEN,
@@ -786,10 +624,13 @@ describe("SPAClient.ts Class Unit Tests", () => {
                     };
                     defaultAuthConfig.storageInterface.setItem(JSON.stringify(accessTokenKey1), JSON.stringify(atValue));
                     sinon.stub(Authority.prototype, <any>"discoverEndpoints").resolves(DEFAULT_OPENID_CONFIG_RESPONSE);
-                    const tokenRequest: TokenRenewParameters = {
-                        scopes: [testScope1]
+                    const tokenRequest: AuthorizationUrlRequest = {
+						redirectUri: TEST_URIS.TEST_REDIR_URI,
+						scopes: [testScope1],
+						codeChallenge: TEST_CONFIG.TEST_CHALLENGE,
+						codeChallengeMethod: Constants.S256_CODE_CHALLENGE_METHOD
                     };
-                    const tokenResponse = await authClient.getValidToken(tokenRequest);
+                    const tokenResponse = await authClient.getValidToken(tokenRequest, null, false);
                     expect(tokenResponse.uniqueId).to.be.empty;
                     expect(tokenResponse.tenantId).to.be.empty;
                     expect(tokenResponse.scopes).to.be.deep.eq([testScope1, Constants.OFFLINE_ACCESS_SCOPE]);
@@ -840,8 +681,7 @@ describe("SPAClient.ts Class Unit Tests", () => {
                         clientId: TEST_CONFIG.MSAL_CLIENT_ID,
                         scopes: testScopes.join(" "),
                         authority: `${Constants.DEFAULT_AUTHORITY}/`,
-                        homeAccountIdentifier: TEST_DATA_CLIENT_INFO.TEST_HOME_ACCOUNT_ID,
-                        resource: "Resource1"
+                        homeAccountIdentifier: TEST_DATA_CLIENT_INFO.TEST_HOME_ACCOUNT_ID
                     };
                     const atValue: AccessTokenValue = {
                         accessToken: TEST_TOKENS.ACCESS_TOKEN,
@@ -854,11 +694,14 @@ describe("SPAClient.ts Class Unit Tests", () => {
                     defaultAuthConfig.storageInterface.setItem(PersistentCacheKeys.ID_TOKEN, TEST_TOKENS.IDTOKEN_V2);
                     defaultAuthConfig.storageInterface.setItem(PersistentCacheKeys.CLIENT_INFO, TEST_DATA_CLIENT_INFO.TEST_RAW_CLIENT_INFO);
                     defaultAuthConfig.storageInterface.setItem(JSON.stringify(accessTokenKey1), JSON.stringify(atValue));
-                    sinon.stub(Authority.prototype, <any>"discoverEndpoints").resolves(DEFAULT_OPENID_CONFIG_RESPONSE);
-                    const tokenRequest: TokenRenewParameters = {
-                        scopes: [testScopes[0]]
+					sinon.stub(Authority.prototype, <any>"discoverEndpoints").resolves(DEFAULT_OPENID_CONFIG_RESPONSE);
+					const tokenRequest: AuthorizationUrlRequest = {
+						redirectUri: TEST_URIS.TEST_REDIR_URI,
+						scopes: [testScopes[0]],
+						codeChallenge: TEST_CONFIG.TEST_CHALLENGE,
+						codeChallengeMethod: Constants.S256_CODE_CHALLENGE_METHOD
                     };
-                    const tokenResponse = await authClient.getValidToken(tokenRequest);
+                    const tokenResponse = await authClient.getValidToken(tokenRequest, null, false);
 
                     // Build Test account
                     const idToken = new IdToken(TEST_TOKENS.IDTOKEN_V2, defaultAuthConfig.cryptoInterface);
@@ -893,7 +736,6 @@ describe("SPAClient.ts Class Unit Tests", () => {
             });
 
             it("returns valid server code response", () => {
-                defaultAuthConfig.storageInterface.setItem(TemporaryCacheKeys.REQUEST_STATE, RANDOM_TEST_GUID);
                 const testSuccessHash = `#code=thisIsATestCode&client_info=${TEST_DATA_CLIENT_INFO.TEST_RAW_CLIENT_INFO}&state=${RANDOM_TEST_GUID}`;
                 defaultAuthConfig.cryptoInterface.base64Decode = (input: string): string => {
                     switch (input) {
@@ -914,20 +756,17 @@ describe("SPAClient.ts Class Unit Tests", () => {
                     }
                 };
                 authModule = new SPAClient(defaultAuthConfig);
-                const codeResponse = authModule.handleFragmentResponse(testSuccessHash);
-                expect(codeResponse.code).to.be.eq(`thisIsATestCode`);
-                expect(codeResponse.userRequestState).to.be.eq(RANDOM_TEST_GUID);
+                const code = authModule.handleFragmentResponse(testSuccessHash, RANDOM_TEST_GUID);
+                expect(code).to.be.eq(`thisIsATestCode`);
             });
 
             it("throws server error when error is in hash", () => {
                 const testErrorHash = `#error=error_code&error_description=msal+error+description&state=${RANDOM_TEST_GUID}`;
 
-                defaultAuthConfig.storageInterface.setItem(TemporaryCacheKeys.REQUEST_STATE, RANDOM_TEST_GUID);
-                expect(() => authModule.handleFragmentResponse(testErrorHash)).to.throw("msal error description");
+                expect(() => authModule.handleFragmentResponse(testErrorHash, RANDOM_TEST_GUID)).to.throw("msal error description");
                 expect(store).to.be.empty;
 
-                defaultAuthConfig.storageInterface.setItem(TemporaryCacheKeys.REQUEST_STATE, RANDOM_TEST_GUID);
-                expect(() => authModule.handleFragmentResponse(testErrorHash)).to.throw(ServerError);
+                expect(() => authModule.handleFragmentResponse(testErrorHash, RANDOM_TEST_GUID)).to.throw(ServerError);
                 expect(store).to.be.empty;
             });
         });
@@ -941,12 +780,14 @@ describe("SPAClient.ts Class Unit Tests", () => {
 
         const postLogoutRedirectUriFunc = () => {
             return TEST_URIS.TEST_LOGOUT_URI;
-        };
+		};
+		
+		const testAuthority = new AadAuthority(TEST_CONFIG.validAuthority, null);
 
         const Client_functionRedirectUris = new SPAClient({
             authOptions: {
                 clientId: TEST_CONFIG.MSAL_CLIENT_ID,
-                authority: TEST_CONFIG.validAuthority,
+                authority: testAuthority,
                 redirectUri: redirectUriFunc,
                 postLogoutRedirectUri: postLogoutRedirectUriFunc,
             },
@@ -992,7 +833,7 @@ describe("SPAClient.ts Class Unit Tests", () => {
         const Client_noRedirectUris = new SPAClient({
             authOptions: {
                 clientId: TEST_CONFIG.MSAL_CLIENT_ID,
-                authority: TEST_CONFIG.validAuthority,
+                authority: testAuthority,
             },
             storageInterface: {
                 getCache(): InMemoryCache {
