--- lib/msal-common/test/response/AuthResponse.spec.ts
@@ -1,12 +1,0 @@
-import { expect } from "chai";
-import { AuthResponse, buildResponseStateOnly } from "../../src/response/AuthResponse";
-import { TEST_CONFIG } from "../utils/StringConstants";
-
-describe("AuthResponse.ts Unit Tests", () => {
-
-    it("buildResponseState creates a new response with only state parameter", () => {
-        let stateOnlyResponse: AuthResponse = buildResponseStateOnly(TEST_CONFIG.STATE);
-        expect(stateOnlyResponse).to.be.not.null;
-        expect(stateOnlyResponse.userRequestState).to.be.eq(TEST_CONFIG.STATE);
-    });
-});
